SiFive: Add PROVIDE() to __global_pointer$ and the relatived symbols.
Theoretically, the compact shared libraries should have their own
__global_pointer$, so their text and data sections can be far apart.
But it will cause the ld "Build pr26590 (1)" testcase fails, and
seems to invalidate the --as-needed option.  Therefore, maybe we
shouldn't generate the __global_pointer$ for the shared libraries,
since it is rarely necessary to put their text and data far away.
However, I notice that add the PROVIDE() to __global_pointer$ can
fix the testcase for medlow and medany code models.  Even if it is
still failed for compact, I think we can consider this issue later.